🐛 Fix useSearchParams
